Update run_metroman.py
When the data are inadequate for MetroMan to run on, it writes out an empty output file. But it currently writes out after removing "bad" data and before adding empty placeholders back into timeseries, which causes problems.

This one line changes DAll.nt so that it will create empty timeseries that are the correct length, when writing empty output files.